// src/Redux/ProductReducer/ProductReducer.types.ts
export type Product = {
  id: 'topcloth_1';
  type: 'topclothes';
  name: 'Top Cloth 1';
  desc: 'Lorem ipsum dolor, sit amet consectetur adipisicing elit. Aliquid, nulla.';
  imgSrc_jpg: './images/clothes/topcloth1_show.jpg';
  imgSrc_png: './images/clothes/topcloth1.png';
};

export type ProductInitialState = Product[];